Tool

+++ Tool/Z.Tool.InternGen/Entry.cs
@@ -2,12 +2,12 @@ namespace Z.Tool.InternGen;
 
 class Entry : EntryEntry
 {
-    protected override int ExecuteMain()
+    protected override long ExecuteMain()
     {
         Gen gen;
         gen = new Gen();
         gen.Init();
-        int o;
+        long o;
         o = gen.Execute();
         return o;
     }
